refactor(context): 🤖 rename stream context builder
Mechanical rename only; behavior is unchanged.
